declare time_corrected as extern and remove antilag_list_players
time_corrected is a global variable declared in g_local.h and defined
in antilag.c. time_corrected is referenced by other translation units
by including the header. antilag_list_players is unused in client.c
and is a global variable in antilag.c.

This resolves linking errors with duplicate symbols on OpenBSD.